pyproject.toml
[project]
name = "mcp-server-collector"
version = "0.1.0"
description = "A MCP Server used to collect MCP Servers over the internet."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "aiohttp>=3.11.10",
 "mcp>=1.1.0",
 "openai>=1.57.0",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
]
[[project.authors]]
name = "idoubi"
email = "me@idoubi.cc"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-collector = "mcp_server_collector:main"
